fix: address property-editor review feedback (ascii folding, shared wrappers, internal-key predicate)
