+++ crates/outl-frontend-shared/CLAUDE.md
@@ -127,7 +127,7 @@ When in doubt, ship in the client; promote later when the second client appears.
 | `isImagePath(href)` → `boolean` (extension classifier — strips `?`/`#`, matches the last `.`-segment; drives whether `<MarkdownInline />`'s `image` token renders an `<img>` or a click-to-open file chip) + `assetFileName(href)` (last non-empty path segment, for the chip label) + `isSafeHttpUrl(href)` → `boolean` (`http`/`https` only; gates whether a remote image href is loaded directly as `<img src>` or refused to a chip, since the href comes from synced/imported markdown) | `@outl/shared/links` | `isImagePath` mirrors `IMAGE_EXTENSIONS` in `crates/outl-md/src/wikilink.rs` (`is_image_target`) — keep the extension list in sync |
 | `formatNextFire(iso, now?)` → `"now"` / `"in 20min"` / `"in 3h"` / `"tomorrow 09:00"` / `"Dec 15, 10:00"`, and `groupReminders(list, now?)` → Today / Tomorrow / This week / Later / Done (empty buckets dropped). Both GUI clients render the same column and the same section headers; two implementations drift on the edge cases (exactly 60 minutes, midnight rollover, a finished rule) long before anyone notices | `@outl/shared/api/commands` | No Rust mirror — presentation only. **The instants themselves come from `outl_actions::reminders::next_fire_at`**, which is the single owner of *when* a reminder fires. Never re-derive a schedule in TS |
 | `Reminder` / `ReminderSettings` DTOs. Times are ISO-8601 **local** strings with no zone suffix, deliberately: the backend resolved them in the user's configured timezone (`outl_actions::clock`), and re-deriving a local time from an epoch in JS reintroduces exactly the bug that module exists to fix | `@outl/shared/api/types` | `outl_tauri_shared::commands::reminders::{ReminderDto, ReminderSettingsDto}` |
-| `<BlockProperties />` — a block's `key:: value` chips under its text, editable in place when `onCommit` is wired (click a chip → `<input>` → Enter/blur commits, Esc reverts, empty clears). Both clients rendered **nothing** here before, so a `remind::` written by a chord left the block looking untouched — invisible and uneditable from the outline. The inline editor is shared rather than per-client because an `<input>` is an input whether a keyboard or the iOS one drives it; only the theme classes differ. `propertyChips(props)` is the pure policy underneath: which keys get a glyph (`⏰ remind`, `▶ auto-run`, `📋 template`), which are the user's own (`priority: high`, rendered verbatim — interpreting it isn't ours), and which are outl's bookkeeping (`id`, `from-template` — hidden). `remindRule(props)` is the "does this block nag me" shortcut | `@outl/shared/markdown` | glyph table mirrored by `outl-tui`'s `property_glyph` (`view/outline.rs`) — a const can't cross the Rust/TS boundary, so the two are edited together |
+| `<BlockProperties />` — a block's `key:: value` chips under its text, editable in place when `onCommit` is wired (click a chip → `<input>` → Enter/blur commits, Esc reverts, empty clears). Both clients rendered **nothing** here before, so a `remind::` written by a chord left the block looking untouched — invisible and uneditable from the outline. The inline editor is shared rather than per-client because an `<input>` is an input whether a keyboard or the iOS one drives it; only the theme classes differ. `propertyChips(props)` is the pure policy underneath: which keys get a glyph (`⏰ remind`, `▶ auto-run`, `📋 template`), which are the user's own (`priority: high`, rendered verbatim — interpreting it isn't ours), and which are outl's bookkeeping (`id`, `from-template` — hidden). `remindRule(props)` is the "does this block nag me" shortcut. `isInternalKey(key)` exports the bookkeeping-key predicate on its own, so property editors (the mobile sheet's key chips) hide the same keys the chip row hides instead of keeping a copy of the list | `@outl/shared/markdown` | glyph table mirrored by `outl-tui`'s `property_glyph` (`view/outline.rs`) — a const can't cross the Rust/TS boundary, so the two are edited together |
 | `looksLikeOutline` | `@outl/shared/paste` | `outl_actions::paste::looks_like_outline` |
 | `hasMultipleParagraphs` | `@outl/shared/paste` | mirror of `split_paragraphs(...).length > 1` in `outl_actions::paste` — gate that decides whether plain text needs the structured backend path |
 | `htmlToOutlMarkdown` | `@outl/shared/paste` | Rich-clipboard `text/html` → outl markdown via **Turndown**, configured for the outl dialect (`*italic*` not `_italic_`, `**bold**`, `- ` bullets collapsed to 2-space nesting, `~~strike~~`, inline `<img alt>` → its alt text so Slack `:emoji:` survives). No Rust mirror — HTML only reaches the GUI webview clients; the resulting markdown then rides the same `paste_markdown_at` backend path as any paste |
