+++ crates/outl-frontend-shared/src/api/commands.ts
@@ -32,6 +32,7 @@ import type {
   PluginToolbarButton,
   PluginTransformer,
   PluginTransformResult,
+  PropertyKey,
   RegistryItem,
   Reminder,
   ReminderSettings,
@@ -1113,4 +1114,30 @@ export function setBlockProperty(
   value: string,
 ): Promise<PageView> {
   return invoke<PageView>("set_block_property", { pageId, blockId, key, value });
+}
+
+/**
+ * Property keys used anywhere in the workspace, most-used first.
+ *
+ * Deliberately **not** cached: the backend answer is a scan of the
+ * property map (no tree walk, no block text), so it is cheaper to ask
+ * every time an editor opens than to hold a list that goes stale the
+ * first time the user adds a key.
+ */
+export function knownPropertyKeys(): Promise<PropertyKey[]> {
+  return invoke<PropertyKey[]>("known_property_keys");
+}
+
+/**
+ * Set — or, with an empty `value`, clear — a property on the **page**
+ * itself (`icon::`, `type::`, …). Rejects the structural keys
+ * (`page-slug`, `page-kind`): those are the page's identity and
+ * renaming is `page_rename`, not a property edit.
+ */
+export function setPageProperty(
+  pageId: string,
+  key: string,
+  value: string,
+): Promise<PageView> {
+  return invoke<PageView>("set_page_property", { pageId, key, value });
 }
