+++ crates/outl-frontend-shared/src/markdown/index.ts
@@ -8,4 +8,9 @@ export {
 } from "./quote";
 export { QuoteWrap, isBlockQuoted, type QuoteWrapProps } from "./QuoteWrap";
 export { BlockProperties, type BlockPropertiesProps } from "./BlockProperties";
-export { propertyChips, remindRule, type PropertyChip } from "./properties";
+export {
+  isInternalKey,
+  propertyChips,
+  remindRule,
+  type PropertyChip,
+} from "./properties";
