+++ crates/outl-mobile/src/lib/api.ts
@@ -38,45 +38,13 @@ export function setWorkspace(path: string): Promise<void> {
 // — both clients register identical commands.
 
 // ── Properties (issue #13) ──────────────────────────────────────────
-// PROMOTE-TO-SHARED: `knownPropertyKeys` / `setPageProperty` wrap
-// commands **both** GUI clients register (`known_property_keys`,
-// `set_page_property`), so by the repo's own rule they belong next to
-// `setBlockProperty` in `@outl/shared/api/commands`. They sit here only
-// because the desktop half of #13 was landing in that file at the same
-// time; move them up on the first commit that owns
-// both sides.
-
-// `PropertyKey` is a wire type both GUI clients share — it lives in
-// `@outl/shared/api/types`, re-exported here so existing imports from
-// this module keep resolving.
+// `knownPropertyKeys` / `setPageProperty` live in
+// `@outl/shared/api/commands` next to `setBlockProperty` — both GUI
+// clients register the same commands, so the wrappers have one owner.
+// Re-exported here (with the `PropertyKey` wire type) so existing
+// imports from this module keep resolving.
 export type { PropertyKey } from "@outl/shared/api/types";
-
-/**
- * Property keys already used somewhere in the workspace, most-used
- * first. The mobile Properties sheet paints the top ones as tappable
- * chips: keys repeat (a graph has a dozen, not a hundred) and typing
- * `oura-date` on a phone keyboard is the cost this removes.
- *
- * Cheap enough to call when the sheet opens — a scan of the property
- * map, no tree walk — so it is never cached into staleness.
- */
-export function knownPropertyKeys(): Promise<PropertyKey[]> {
-  return invoke<PropertyKey[]>("known_property_keys");
-}
-
-/**
- * Set — or clear, with an empty `value` — a `key:: value` property on
- * the **page** itself (`icon::`, `type::`, …). Returns the refreshed
- * page view.
- *
- * The backend refuses the structural keys (`page-slug`, `page-kind`):
- * they are the page's identity, and renaming is `page_rename`, which
- * moves the projection too.
- */
-export function setPageProperty(
-  pageId: string,
-  key: string,
-  value: string,
-): Promise<import("@outl/shared/api/types").PageView> {
-  return invoke("set_page_property", { pageId, key, value });
-}
+export {
+  knownPropertyKeys,
+  setPageProperty,
+} from "@outl/shared/api/commands";
